Compute simd mul table using `const fn` instead of using MACROs
